Split refresh panels
